==> StrategicNonCitations/4b-InventorMobilityNewCites-ControlSim-0927.ipynb <==
import numpy as np
import scipy as sp
import pandas as pd
import timeit
import datetime
import time
import pprint
import itertools
import pickle
import sklearn
import dask
import os
os.chdir('/mnt/t48/bighomes-active/sfeng/patentdiffusion/')
import fastparquet
seed = 3
import seaborn as sns
import matplotlib
import matplotlib.pyplot as plt
import h5py
import dask
import dask.dataframe as dd
import dask.array as da
# Percentiles
from scipy.stats import percentileofscore
from IPython.display import display

For sample generation, see: https://sfengc7.stern.nyu.edu:8888/notebooks/patentdiffusion/201808Results/StrategicNonCitations/Previous/1d-InventorMobilityNewCitations-0911.ipynb

See previous notebook for similarity to cited patent:
https://sfengc7.stern.nyu.edu:8888/notebooks/patentdiffusion/201808Results/StrategicNonCitations/4a-InventorMobilityNewCites-0918.ipynb

yv = "appyear"
c2 = pd.read_pickle("DataStore/2018-08/inv_move_cites_0912.pkl")
ip = pd.read_pickle("DataStore/2018-08/inv_move_pats_0912.pkl")

# Use unique assignees
c2 = c2.drop(["cited_asg", "citing_asg"],1)
asgs = fastparquet.ParquetFile("RawData/Cleaned/patent_assignees_unique_0628.parq").to_pandas(["patent", "assignee_id"])
pdf = fastparquet.ParquetFile("RawData/Cleaned/patent_loc_unique_us_0628.parq")\
.to_pandas(["patent", "primclass", "appyear"])
pdf = pdf.merge(asgs, how = "left", on = "patent")

c2 = c2.merge(asgs, how="left", left_on="citing", right_on="patent").rename(columns={"assignee_id": "citing_asg"}).drop("patent",1)
c2 = c2.merge(asgs, how="left", left_on="cited", right_on="patent").rename(columns={"assignee_id": "cited_asg"}).drop("patent",1)
del(asgs)

# New firms that cite prior patent post move
a1 = c2.loc[(c2["citing_appyear"] < c2["sec_fyear"]), "citing_asg"].tolist()
a2 = c2.loc[(c2["citing_appyear"] >= c2["sec_fyear"]), "citing_asg"].tolist()
new_cite_asgs = list(set(a2).difference(set(a1)))

c3 = pd.read_pickle("DataStore/2018-08/inv_mob_cite_pc_control_0918.pkl")

# Patents by newly citing assignees
pdf = pdf.loc[pdf["assignee_id"].isin(new_cite_asgs)]
len(pdf)

# Sort by assignee, primclass, app year
pdf = pdf.sort_values(["assignee_id", "primclass", yv], ascending = [1,1,0])

# Control patent by assignee, primclass
cdict = {}
adict = {}
for yr in range(1975, 2016):
    print(yr)
    print(datetime.datetime.now())
    # Patent by assignee, primclass
    p2 = pdf.loc[(pdf["appyear"].isin(range(yr-5,yr+1))), \
        ["appyear", "assignee_id", "primclass", "patent"]].groupby(["assignee_id", "primclass"])
    p2 = {n+(yr,): (g["patent"].tolist() if len(g["patent"].tolist()) >= 1 else None) for n,g in p2}
    cdict.update(p2)
    del(p2)
    
     # Patent by assignee
    p2 = pdf.loc[(pdf["appyear"].isin(range(yr-5,yr+1))), \
        ["appyear", "assignee_id", "primclass", "patent"]].groupby(["assignee_id"])
    p2 = {(n,yr): (g["patent"].tolist() if len(g["patent"].tolist()) >= 1 else None) for n,g in p2}
    adict.update(p2)
    del(p2)

# Get all of cited patent's citations
c2 = pd.read_pickle("DataStore/2018-08/inv_move_cites_0912.pkl")
c2 = {n:g["citing"].tolist() for n,g in c2[["cited", "citing"]].groupby("cited")}

# Get all firm's patents
%time a = [adict.get((asg, fyr), []) for asg,fyr in zip(c3["citing_asg"], c3["citing_appyear"])]
c3["asg_pats"] = a

# Get list of potential control candidates
%time c = [cdict.get((asg, pc, fyr), []) for asg,pc,fyr in zip(c3["citing_asg"], c3["citing_primclass"],\
                                                     c3["citing_appyear"])]
# Add to dataframe
c3["asg_pc_pats"] = c

c3.to_pickle("DataStore/2018-08/inv_mob_cite_pc_control_0927.pkl")

### Create similarity pair sample

#### First, find how long each pairwise list would be for (i) all firm's patents in previous 5 years; (ii) all firm's primary class patents

# Use itertools to get pairwise list
apats = (itertools.product(l,l) for l in c3["asg_pats"])
apats = [item for sublist in apats for item in sublist]
print(len(apats))

ppats = (itertools.product(l,l) for l in c3["asg_pc_pats"])
ppats = [item for sublist in ppats for item in sublist]
print(len(ppats))

del(apats, ppats)

del(apats, ppats)

#### Create pairs for the citing patent and control patent, each crossed with all firm's patents in in previous five years

# c4 = pd.DataFrame()

# for c in ["citing", "citing_control_asg_pc"]:
#     apats = (itertools.product([i],l) for i,l in zip(c3[c], c3["asg_pats"]))
#     apats = [item for sublist in apats for item in sublist]
#     apats = pd.DataFrame({"tp": [i[0] for i in apats], "op": [i[1] for i in apats]})
#     apats["type"] = c
#     c4 = c4.append(apats, ignore_index=True)
#     del(apats)
# print(len(c4))

# # Delete duplicates and self similarity
# c4 = c4.loc[~(c4["tp"] == c4["op"])]
# c4 = c4.drop_duplicates()

# Add original citing
co = dict(zip(c3["citing"], c3["citing"]))
co.update(dict(zip(c3["citing_control_asg_pc"], c3["citing"])))
c4["citing"] = c4["tp"].map(co)

def grouper(n, iterable):
    """
    >>> list(grouper(3, 'ABCDEFG'))
    [['A', 'B', 'C'], ['D', 'E', 'F'], ['G']]
    """
    iterable = iter(iterable)
    return iter(lambda: list(itertools.islice(iterable, n)), [])


import scipy.spatial.distance as distance
dms = ["ldavecs", "docvecs"]

print("Getting row values")
print(datetime.datetime.now())
pat_dict = fastparquet.ParquetFile("RawData/Cleaned/patabs7615_us_no_dup.parq").to_pandas(["patent"])["patent"].tolist()
pat_dict = dict(zip(pat_dict, range(len(pat_dict))))


l2 = c4.copy()

for dm in dms:
    print((dm,"started"))
    print("Loading matrix and dict")
    print(datetime.datetime.now())
    
    if dm == "ldavecs":
        ncols = 60
    else:
        ncols = 100

    pm = fastparquet.ParquetFile("DataStore/2018-07-P2/ML/{0}_pats_0712.parq".format(dm))\
    .to_pandas().values
    
    
    l3 = l2[["tp", "op"]]
    l3 = l3.dropna(how="any").drop_duplicates()
    # Store copy as array
    l3 = l3.loc[l3["tp"].isin(pat_dict.keys()) & l3["op"].isin(pat_dict.keys())]
    print(len(l3))

    # Convert to chunks
    print("Getting chunks")
    print(datetime.datetime.now())
    # Split into chunks
    n_rows = 3000
    n_chunks = int(np.round(len(l3)/n_rows))
    tp_chunks = grouper(n_rows, pm[[pat_dict[p[1]] for p in l3["tp"].iteritems()]])
    op_chunks = grouper(n_rows, pm[[pat_dict[p[1]] for p in l3["op"].iteritems()]])

    chunks = itertools.zip_longest(tp_chunks, op_chunks)

    print("Getting patent pair cosine similarity")
    print(datetime.datetime.now())
    # Cosine

    cos_dis = np.empty(len(l3))

    for r, c in enumerate(chunks):
        cos_dis[r*n_rows:r*n_rows+n_rows] = np.diag(distance.cdist(c[0],c[1], metric = "cosine"))

    l3["sim_{0}".format(dm)] = 1-cos_dis

    # Rename columns
    l2 = l2.merge(l3, how = "left", on = ["tp", "op"])
    print(len(l2))
    del(l3)
    print("finished")
    print(datetime.datetime.now())
    del(pm)

c4 = l2
c4 = c4.loc[~(c4["tp"] == c4["op"])]
c4 = c4.drop_duplicates()

# Add primary class of citing and other
pdf = fastparquet.ParquetFile("RawData/Cleaned/patent_loc_unique_us_0628.parq")\
.to_pandas(["patent", "primclass"])
pdf = dict(zip(pdf["patent"], pdf["primclass"]))
c4["citing_primclass"] = c4["citing"].map(pdf)
c4["op_primclass"] = c4["op"].map(pdf)
c4["citing_primclass_match"] = (c4["citing_primclass"] == c4["op_primclass"])
print(len(c4))
c4.to_pickle("DataStore/2018-08/inv_mobility_cite_sim_0927.pkl")

acite = c4.loc[(c4["type"] == "citing"), ["citing", "sim_ldavecs", "sim_docvecs"]].groupby("citing").mean()\
.rename(columns={"sim_ldavecs": "citing_mean_asg_ldavecs", "sim_docvecs": "citing_mean_asg_docvecs"})
acont = c4.loc[(c4["type"] == "citing_control_asg_pc"), ["citing", "sim_ldavecs", "sim_docvecs"]].groupby("citing").mean()\
.rename(columns={"sim_ldavecs": "control_mean_asg_ldavecs", "sim_docvecs": "control_mean_asg_docvecs"})
pcite = c4.loc[(c4["type"] == "citing") & (c4["citing_primclass_match"] == True),
               ["citing", "sim_ldavecs", "sim_docvecs"]].groupby("citing").mean()\
.rename(columns={"sim_ldavecs": "citing_mean_asg_pc_ldavecs", "sim_docvecs": "citing_mean_asg_pc_docvecs"})
pcont = c4.loc[(c4["type"] == "citing_control_asg_pc") & (c4["citing_primclass_match"] == True),
               ["citing", "sim_ldavecs", "sim_docvecs"]].groupby("citing").mean()\
.rename(columns={"sim_ldavecs": "control_mean_asg_pc_ldavecs", "sim_docvecs": "control_mean_asg_pc_docvecs"})


c5 = pd.concat([acite, acont, pcite, pcont], axis=1).reset_index()

c5.dropna(how="any").describe()

c3 = c3.merge(c5, how="left", on="citing")

c3 = c3.rename(columns={"sim_ldavecs": "citing_ldavecs", "sim_docvecs": "citing_docvecs",
                       "sim_ldavecs_citing_control_asg_pc": "control_ldavecs",
                        'sim_docvecs_citing_control_asg_pc': "control_docvecs"})

c3.to_pickle("DataStore/2018-08/inv_mob_cite_pc_control_0927.pkl")

### Analysis: Create data tables

dm = "docvecs"
tab = pd.DataFrame({})
for c in [dm, "mean_asg_{0}".format(dm), "mean_asg_pc_{0}".format(dm)]:
    c4 = c3.dropna(subset=["citing_{0}".format(c), "control_{0}".format(c)], how="any")
    cite_m = c4["citing_{0}".format(c)].mean()
    cont_m = c4["control_{0}".format(c)].mean()
    # Independent samples
#     t1 = sp.stats.ttest_ind(c3["citing_{0}".format(c)], c3["control_{0}".format(c)], equal_var=False, nan_policy="omit")
    # Related samples
    t2 = sp.stats.ttest_rel(c4["citing_{0}".format(c)], c4["control_{0}".format(c)], nan_policy="omit")
    tab[c] = [cite_m, cont_m, t2[0], t2[1], len(c4)]
    
tab.columns = ["Sim DocVecs to Cited", "\makecell{Mean Sim Docvecs,\\\\Own Prior Pats}",
               "\makecell{Mean Sim Docvecs,\\\\Own Prior Pats in Citing PC}"]
tab.index = ["Citing", "Control", "$t$-value", "$p$-value", "$N$"]
 

tab2 = np.round(tab,3)
tab2.loc["$N$"] = tab2.loc["$N$"].astype(int).astype(str)
print(tab2.to_latex(escape=False,column_format="lccc"))

### Stats on inventors moving

yv = "appyear"
# All inventors who have moved
ip = pd.read_pickle("DataStore/2018-07/inv_move_pats_0712.pkl")
print(len(ip))

pdf = fastparquet.ParquetFile("RawData/Cleaned/patent_loc_unique_us_0628.parq")\
.to_pandas(["patent", "inv_msa", "gyear", "appyear"])

# Add application year
ip[yv] = ip["patent"].map(dict(zip(pdf["patent"], pdf[yv])))

# Sort by inventor, grant year
ip = ip.sort_values(["inventor_id", yv])

# Only look at inventors' first and second cities
ip = ip.loc[(ip["inv_asg_rank"] <= 1)]

# Inventors' second cities
sc = ip.loc[(ip["inv_asg_rank"] == 1), ["inventor_id", "inv_msa", yv]].drop_duplicates(["inventor_id", "inv_msa"])

# Inventors' second city compared to first
ip["sec_inv_msa"] = ip["inventor_id"].map(dict(zip(sc["inventor_id"], sc["inv_msa"])))

# Second city's first grant year
ip["sec_fyear"] = ip["inventor_id"].map(dict(zip(sc["inventor_id"], sc[yv])))

# # Get rid of the inventors whose second MSA matches the first
# ip = ip.loc[~(ip["inv_msa"] == ip["sec_inv_msa"])]
# print(len(ip))

print(len(ip.loc[~(ip["inv_msa"] == ip["sec_inv_msa"])]), len(ip.loc[(ip["inv_msa"] == ip["sec_inv_msa"])]),
      len(ip), len(ip.loc[~(ip["inv_msa"] == ip["sec_inv_msa"])])/len(ip))

### Mobile inventors' prior patent citations

ip = pd.read_pickle("DataStore/2018-08/inv_move_pats_0912.pkl")
ip.columns

ip.head()
# Two approaches
# 1. Drop rows with missing values
i2 = ip[['sec_inv_msa_match_prior', 'sec_inv_msa_match_post']].dropna(how="any")
display(i2.describe())

# 2. Only use patents that received citations, i.e. is not null for both prior and post. Then Fill nan with 0.
i3 = ip[['sec_inv_msa_match_prior', 'sec_inv_msa_match_post']].dropna(how="all").fillna(0)
display(i3.describe())

### Number of assignees that cite mobile inventors patents

print(len(set(a1)), len(set(a2)), len(set(new_cite_asgs)), len(set(new_cite_asgs))/len(set(a2)))

# Number of citations from new assignees
print(len(c3))
